feature(entity): add HasOwnerTrait
