+++ src/Entity/Herd.php
@@ -3,6 +3,7 @@
 namespace App\Entity;
 
 use App\Entity\Interface\HasOwner;
+use App\Entity\Trait\HasOwnerTrait;
 use App\Repository\HerdRepository;
 use DateTimeImmutable;
 use Doctrine\ORM\Mapping as ORM;
@@ -12,8 +13,12 @@ use Symfony\Component\Serializer\Attribute\Groups;
 
 #[ORM\Entity(repositoryClass: HerdRepository::class)]
 #[Gedmo\SoftDeleteable(fieldName: 'deletedAt', timeAware: false, hardDelete: false)]
+#[ORM\AssociationOverrides([
+    new ORM\AssociationOverride(name: 'owner', inversedBy: 'herds'),
+])]
 class Herd implements HasOwner
 {
+    use HasOwnerTrait;
     use SoftDeleteableEntity;
 
     #[ORM\Id]
@@ -33,10 +38,6 @@ class Herd implements HasOwner
     #[ORM\Column]
     #[Groups(['herd'])]
     private ?DateTimeImmutable $createdAt = null;
-
-    #[ORM\ManyToOne(inversedBy: 'herds')]
-    #[ORM\JoinColumn(nullable: false)]
-    private ?User $owner = null;
 
     public function getId(): ?int
     {
@@ -75,18 +76,6 @@ class Herd implements HasOwner
     public function setCreatedAt(DateTimeImmutable $createdAt): static
     {
         $this->createdAt = $createdAt;
-
-        return $this;
-    }
-
-    public function getOwner(): ?User
-    {
-        return $this->owner;
-    }
-
-    public function setOwner(?User $owner): static
-    {
-        $this->owner = $owner;
 
         return $this;
     }
